Fix tabulation of files and put all gql queries in seperate folder
