--- client/src/component/index.js
@@ -1,3 +1,0 @@
-import category from './category';
-
-export default category;
