+++ client/src/redux/category/action.js
@@ -1,5 +1,3 @@
-import { gql } from '@apollo/client';
-
 export const GETCATEGORY = 'CATEGORIES/CURRENTCATEGORY/GET';
 
 export const fetchCategory = (payload) => (
@@ -8,34 +6,3 @@ export const fetchCategory = (payload) => (
     payload,
   }
 );
-
-export const CATEGORY_QUERY = (payload) => gql`
-query CategoryQuery{
-  category(input:{title:"${payload}"} ){
-  name,
-  products{
-    id,
-    name,
-    inStock,
-    gallery,
-    prices{
-      currency{
-       symbol,
-       label
-     },
-       amount
-     },
-     attributes{
-      id,
-      name,
-      type,
-      items{
-        displayValue,
-        id,
-        value
-      }
-      }
-  }
-}
-}
-`;
