+++ client/src/redux/details/action.js
@@ -1,4 +1,12 @@
 export const GET_PRODUCT_DATA = 'DATA/PRODUCT/GET';
+export const GETPRODUCT = 'CATEGORY/CURRENTPRODUCT/GET';
+
+export const fetchProduct = (payload) => (
+  {
+    type: GETPRODUCT,
+    payload,
+  }
+);
 
 const productDatafn = (payload) => (
   {
